overloaded flugSuche, reform S.FlugDAO
- overloaded flugSuche in FlugController for seraching with details
- form boolean condition for flugEintragen to act correspondedly to the
actual result
- bring back flugDAO initialization to create a new file on the file
path, if there wasn't any (it is more effective this way to make sure
there is a file. In addition, the leerlist-ArrayList<> would be
unnecessary, which means we save some memory usage. (Because it is the
instance variable, which is not used anymore in other functions as it
might cause problems in the paralellization critical region)
- reform for-loop to foreach-loop for reader-comfort and overall
performance